# api.yml
...
openapi: "3.0.0"
info:
  version: 1.0.0
  title: User Service
  license:
    name: MIT
servers:
  - url: http://localhost
paths:
...
  /profile:
    get:
      operationId: profile
      description: Profile User base on user token
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: Profile Detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetProfileResOk"
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: array
...
    GetProfileResOk:
      description: User profile response OK
      type: object
      required:
        - phone
        - name
      properties:
        phone:
          type: string
        name:
          type: string